Fix & Improve the code, remove bugs

+++ src/utils/cloudinary.js
@@ -3,11 +3,10 @@ import fs from 'fs';
 
 (async function() {
     // Configuration
-    return cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
-    });
+    cloudinary.config({ 
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET });
 })();
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -18,13 +17,12 @@ const uploadOnCloudinary = async (localFilePath) => {
         })
         // file has been uploaded successfully
         // console.log("File uploaded on Cloudinary successfully", response.url);
-        fs.unlinkSync(localFilePath);
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
         return response.url;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath);
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
         return null
-        
     }
 }
